another refactor

+++ src/routes/admin/[quizCode]/+page.server.ts
@@ -1,54 +1,69 @@
 import type { Actions, PageServerLoadEvent, RequestEvent } from "./$types";
-import { and, asc, eq, gt, gte, lt, lte, sql } from "drizzle-orm";
-import { answers_, db, questions_, quizzes_ } from "$lib/server";
+import { and, asc, eq, gt, gte, inArray, lt, lte, sql } from "drizzle-orm";
+import { answers_, db, questions_, quizzes_, users_, type Question, type Quiz } from "$lib/server";
 import { error, redirect } from "@sveltejs/kit";
 import { zfd } from "zod-form-data";
 
 export async function load({ params: { quizCode } }: PageServerLoadEvent) {
   if (!quizCode) redirect(303, "/admin");
 
-  const quiz = await db.query.quizzes_.findFirst({
-    where: eq(quizzes_.code, quizCode),
-    with: {
-      questions: { orderBy: asc(questions_.index), with: { answers: true } },
-      users_bridge: {
-        columns: {},
-        with: { to_user: { with: { answers_bridge: { columns: {}, with: { to_answer: true } } } } },
-      },
-    },
-  });
+  const quiz = await db.query.quizzes_.findFirst({ where: eq(quizzes_.code, quizCode) });
   if (!quiz) error(404, "quiz not found");
+  const questions = await db.query.questions_.findMany({
+    where: inArray(questions_.id, quiz.questionIDs),
+    orderBy: asc(questions_.index),
+  });
+  const answers = await Promise.all(
+    questions.map((q) => db.query.answers_.findMany({ where: inArray(answers_.id, q.answerIDs) })),
+  );
+  const users = await db.query.users_.findMany({ where: inArray(users_.id, quiz.userIDs) });
 
-  return { quiz };
+  return { quiz, questions, answers, users };
 }
 
 export const actions: Actions = {
   async addquestion({ request }: RequestEvent) {
-    const { quizCode } = zfd.formData({ quizCode: zfd.text() }).parse(await request.formData());
+    const { quizCode, lastIndex } = zfd
+      .formData({ quizCode: zfd.text(), lastIndex: zfd.numeric() })
+      .parse(await request.formData());
 
-    const questionsL = (await db.query.quizzes_.findFirst({
-      where: eq(quizzes_.code, quizCode),
-      with: { questions: true },
-    }))!.questions.length;
-
-    await db.insert(questions_).values({ quizCode, index: questionsL + 1 });
+    const question = await db
+      .insert(questions_)
+      .values({ index: lastIndex + 1 })
+      .returning();
+    await db
+      .update(quizzes_)
+      .set({ questionIDs: sql`array_append(${quizzes_.questionIDs}, ${question[0].id})` })
+      .where(eq(quizzes_.code, quizCode));
   },
 
   async delquestion({ request }: RequestEvent) {
-    const { questionID } = zfd
-      .formData({ questionID: zfd.numeric() })
+    const { quizCode, questionID } = zfd
+      .formData({ quizCode: zfd.text(), questionID: zfd.numeric() })
       .parse(await request.formData());
 
-    const question = (await db.query.questions_.findFirst({
-      where: eq(questions_.id, questionID),
-    }))!;
+    const [quiz, question]: [Quiz | undefined, Question | undefined] = await Promise.all([
+      db.query.quizzes_.findFirst({
+        where: eq(quizzes_.code, quizCode),
+      }),
+      db.query.questions_.findFirst({
+        where: eq(questions_.id, questionID),
+      }),
+    ]);
+    if (!quiz || !question) return;
 
-    await db.delete(answers_).where(eq(answers_.questionID, question.id));
-    await db
-      .update(questions_)
-      .set({ index: sql`${questions_.index} - 1` })
-      .where(gt(questions_.index, question.index));
-    await db.delete(questions_).where(eq(questions_.id, questionID));
+    await Promise.all([
+      db.delete(answers_).where(inArray(answers_.id, question.answerIDs)),
+      db
+        .update(questions_)
+        .set({ index: sql`${questions_.index} - 1` })
+        .where(and(gt(questions_.index, question.index), inArray(questions_.id, quiz.questionIDs))),
+      db
+        .update(quizzes_)
+        .set({ questionIDs: sql`array_remove(${quizzes_.questionIDs}, ${questionID})` })
+        .where(and(gt(questions_.index, question.index), inArray(questions_.id, quiz.questionIDs))),
+      db.delete(questions_).where(eq(questions_.id, questionID)),
+    ]);
   },
 
   async addanswer({ request }: RequestEvent) {
@@ -56,18 +71,22 @@ export const actions: Actions = {
       .formData({ questionID: zfd.numeric() })
       .parse(await request.formData());
 
-    const question = (await db.query.questions_.findFirst({
-      where: eq(questions_.id, questionID),
-    }))!;
-
-    await db.insert(answers_).values({ questionID: question.id });
+    const answer = await db.insert(answers_).values({}).returning();
+    await db
+      .update(questions_)
+      .set({ answerIDs: sql`array_append(${questions_.answerIDs}, ${answer[0].id})` })
+      .where(eq(questions_.id, questionID));
   },
 
   async delanswer({ request }: RequestEvent) {
     const { answerID } = zfd.formData({ answerID: zfd.numeric() }).parse(await request.formData());
 
-    await db.delete(answers_).where(eq(answers_.id, answerID));
-    await db.update(questions_).set({ correctID: null }).where(eq(questions_.correctID, answerID));
+    await Promise.all([
+      db
+        .update(questions_)
+        .set({ answerIDs: sql`array_remove(${questions_.answerIDs}, ${answerID})` }),
+      db.delete(answers_).where(eq(answers_.id, answerID)),
+    ]);
   },
 
   async coranswer({ request }: RequestEvent) {
@@ -79,11 +98,11 @@ export const actions: Actions = {
   },
 
   async changequizname({ request }: RequestEvent) {
-    const { quizCode, name } = zfd
-      .formData({ quizCode: zfd.text(), name: zfd.text() })
+    const { quizCode, title } = zfd
+      .formData({ quizCode: zfd.text(), title: zfd.text() })
       .parse(await request.formData());
 
-    await db.update(quizzes_).set({ name }).where(eq(quizzes_.code, quizCode));
+    await db.update(quizzes_).set({ title }).where(eq(quizzes_.code, quizCode));
   },
 
   async changequestiontitle({ request }: RequestEvent) {
@@ -123,16 +142,16 @@ export const actions: Actions = {
   async togglestatus({ request }: RequestEvent) {
     const { quizCode } = zfd.formData({ quizCode: zfd.text() }).parse(await request.formData());
 
-    const quiz = (await db.query.quizzes_.findFirst({
-      where: eq(quizzes_.code, quizCode),
-      with: { questions: true },
-    }))!;
+    const quiz = (await db.query.quizzes_.findFirst({ where: eq(quizzes_.code, quizCode) }))!;
     if (quiz.status >= 0) {
       await db.update(quizzes_).set({ status: -1 }).where(eq(quizzes_.code, quizCode));
       return;
     }
 
-    if (quiz.questions.some((q) => !q.correctID))
+    const questions = (await db.query.questions_.findMany({
+      where: inArray(questions_.id, quiz.questionIDs),
+    }))!;
+    if (questions.some((q) => !q.correctID))
       return { error: "choose correct answers for each question!" };
 
     await db.update(quizzes_).set({ status: 0 }).where(eq(quizzes_.code, quizCode));
